tests/wpt/web-platform-tests/clipboard-apis/async-write-duplicate-mime-type-manual.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>
  Async Clipboard write duplicate mime type test
</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
promise_test(async t => {
  const blobText = new Blob(["test text"], {type: 'text/plain'});
  const blobText2 = new Blob(["test text"], {type: 'text/plain'});

  assert_equals(blobText.type, "text/plain");
  assert_equals(blobText2.type, "text/plain");


  await promise_rejects(t, 'NotAllowedError',
      navigator.clipboard.write([blobText, blobText2]));
}, "Verify write and read clipboard (multiple blobs)");
</script>
<p>
  Note: This is a manual test because it writes/reads to the shared system
  clipboard and thus cannot be run async with other tests that might interact
  with the clipboard.
</p>
